Title: 우리 지금 만나 당장 만나, Time: 133ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240228/우리 지금 만나 당장 만나/meet-now-right-now.cpp
@@ -33,7 +33,7 @@ int main() {
     double ans;
 
     for (int i = 0; i < 100; i++) {
-        long long mid = (l + r) / 2;
+        double mid = (l + r) / 2;
 
         if (IsPossible(mid)) {
             r = mid;
